requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.5.0
    # via httpcore
certifi==2021.10.8
    # via
    #   httpcore
    #   httpx
charset-normalizer==2.0.12
    # via httpx
click==8.1.2
    # via
    #   click-default-group
    #   sqlite-utils
click-default-group==1.2.2
    # via sqlite-utils
gitdb==4.0.9
    # via gitpython
gitpython==3.1.27
    # via -r requirements.in
h11==0.12.0
    # via httpcore
httpcore==0.14.7
    # via httpx
httpx==0.22.0
    # via -r requirements.in
idna==3.3
    # via
    #   anyio
    #   rfc3986
jinja2==3.1.1
    # via -r requirements.in
markupsafe==2.1.1
    # via jinja2
python-dateutil==2.8.2
    # via sqlite-utils
rfc3986[idna2008]==1.5.0
    # via httpx
six==1.16.0
    # via python-dateutil
smmap==5.0.0
    # via gitdb
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlite-fts4==1.0.1
    # via sqlite-utils
sqlite-utils==3.25.1
    # via -r requirements.in
tabulate==0.8.9
    # via sqlite-utils
